tuna: bluetooth: deassert BT_REG and BT_NRST during boot
BT_REG and BT_NRST should be deasserted during boot so that they
remain low and they are pulled high using the rfkill switch. Having
the BT_NRST gpio high would power on the chip even when BT was not
in use after boot and would need a power on and off cycle to disable
the chip.

2 Minute silence for all the power that has been lost since November 2011.

Change-Id: I43d2f89dfdcede09adb4ebb7d7378ccaa5ccab1e